Move metrics tracker from shared to plugin reporting
