--- src/shared/reporting.js
@@ -1,1 +1,0 @@
-export { createMetricsTracker } from "./reporting/metrics-tracker.js";
